Removed SquareImageView and replaced it with simple ImageView.
