--- app/src/main/java/org/imozerov/streetartview/ui/detail/SquareImageView.kt
@@ -1,21 +1,0 @@
-package org.imozerov.streetartview.ui.detail
-
-import android.content.Context
-import android.util.AttributeSet
-import android.widget.ImageView
-
-/**
- * Created by sergei on 11.02.16.
- */
-class SquareImageView : ImageView {
-    constructor(context: Context) : super(context)
-    constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
-    constructor(context: Context, attrs: AttributeSet, defStyleAttr: Int) : super(context, attrs, defStyleAttr)
-
-    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec)
-
-        val width = measuredWidth
-        setMeasuredDimension(width, width)
-    }
-}
