+++ app/src/main/res/layout/art_object_gallery_item.xml
@@ -4,7 +4,7 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical">
 
-    <org.imozerov.streetartview.ui.detail.SquareImageView
+    <ImageView
         android:id="@+id/image_preview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
